models/semseg/requirements.txt
astor==0.8.1
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
contextlib2==0.6.0.post1
coverage==5.3
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
future==0.18.2
hyperopt==0.1.2
idna==2.10
imageio==2.9.0
joblib==0.17.0
json-tricks==3.15.4
kiwisolver==1.3.1
matplotlib==3.3.3
netifaces==0.10.9
networkx==2.5
nni==1.9
numpy==1.19.4
Pillow==8.0.1
pkginfo==1.6.1
protobuf==3.14.0
psutil==5.7.3
pymongo==3.11.1
pyparsing==2.4.7
python-dateutil==2.8.1
PythonWebHDFS==0.2.3
PyWavelets==1.1.1
requests==2.25.0
responses==0.12.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
schema==0.7.2
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
simplejson==3.17.2
six==1.15.0
tensorboardX==2.1
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.7.0
torchvision==0.8.1
tqdm==4.53.0
typing-extensions==3.7.4.3
urllib3==1.26.2
websockets==8.1
